+++ app/Lebiru.FileService.AppHost/Program.cs
@@ -3,6 +3,7 @@ var builder = DistributedApplication.CreateBuilder(args);
 builder.AddProject<Projects.Lebiru_FileService>("felix-fileservice", launchProfileName: null)
     .WithHttpEndpoint(port: 3002, name: "http")
     .WithExternalHttpEndpoints()
+    .WithEnvironment("Aspire__DashboardUrl", "http://localhost:18888")
     .WithEnvironment("OpenTelemetry__UseConsoleExporter", "false")
     .WithEnvironment("OTEL_TRACES_SAMPLER", "always_on")
     .WithEnvironment("OTEL_BSP_SCHEDULE_DELAY", "1000")